requirements.txt
numpy
vtk
matplotlib
pypng
